Finally can format

+++ basicFileSystemDisk.h
@@ -5,7 +5,7 @@
 
 class BasicFileSystemDisk : public BasicFileSystem {
 public:
-	int format(const char* name, int size);
+	int format(const char* name, unsigned long size);
 	int read_block(const char* name, int block_index, char *buffer);
 	int write_block(const char* name, int block_index, const char *buffer);
 };
